# fix delete function