- when deleting a branch node, we first walk up the tree to delete
  children. As all siblings and the parents are already being deleted,
  there is no need to walk back down the tree removing the children and
  any empty branches.