@woodsaj 2017-04-17
just added a benchmark.

Old code
```
$ go test -v -run None -bench BenchmarkDeletes -benchtime 10s
BenchmarkDeletes-4   	   50000	   2384248 ns/op	 1313605 B/op	      27 allocs/op
PASS
ok  	github.com/raintank/metrictank/idx/memory	123.452s
```

New Code
```
$ go test -v -run None -bench BenchmarkDeletes -benchtime 10s
BenchmarkDeletes-4   	 5000000	      4145 ns/op	    1354 B/op	       8 allocs/op
PASS
ok  	github.com/raintank/metrictank/idx/memory	66.588s